+++ apps/cold-api/src/platform/modules/resources/organizations/products/dto/create-product.dto.ts
@@ -1,0 +1,4 @@
+export class CreateProductDto {
+  name: string;
+  organization_id: string;
+}